pyproject.toml
[project]
name = "optimizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "pynacl>=1.6.2",
    "python-dotenv>=1.2.1",
    "spotipy>=2.25.2",
    "yt-dlp>=2025.12.8",
]
